Keep picker local to follow action
